+++ BE/src/main/java/roomescape/domain/ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.entity;
+package roomescape.domain;
 
 import java.time.LocalDate;
 import java.util.List;
