+++ BE/src/main/java/roomescape/domain/ReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.entity;
+package roomescape.domain;
 
 import java.util.List;
 import java.util.Optional;
