+++ BE/src/main/java/roomescape/domain/ThemeSortType.java
@@ -1,4 +1,4 @@
-package roomescape.entity;
+package roomescape.domain;
 
 public enum ThemeSortType {
     POPULAR("COUNT(r.id) DESC");
