+++ BE/src/main/java/roomescape/presentation/ReservationController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.application.ReservationService;
-import roomescape.entity.Reservation;
+import roomescape.domain.Reservation;
 import roomescape.presentation.dto.AvailableReservationResponse;
 import roomescape.presentation.dto.ReservationRequest;
 import roomescape.presentation.dto.ReservationResponse;
