+++ BE/src/main/java/roomescape/presentation/dto/ReservationResponse.java
@@ -1,6 +1,6 @@
 package roomescape.presentation.dto;
 
-import roomescape.entity.Reservation;
+import roomescape.domain.Reservation;
 
 public record ReservationResponse(
         Long id,
