+++ BE/src/test/java/roomescape/infrastructure/ReservationTimeJdbcTemplateRepositoryTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
-import roomescape.entity.ReservationTime;
+import roomescape.domain.ReservationTime;
 
 @JdbcTest
 @Import(ReservationTimeJdbcTemplateRepository.class)
