+++ BE/src/test/java/roomescape/infrastructure/ThemeJdbcTemplateRepositoryTest.java
@@ -12,8 +12,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.entity.Theme;
-import roomescape.entity.ThemeSortType;
+import roomescape.domain.Theme;
+import roomescape.domain.ThemeSortType;
 
 @JdbcTest
 @Import(ThemeJdbcTemplateRepository.class)
